Update a lot of things to be pep8, fixed some missing engines from the all flag, introduce pytest for tests, f-stringify some things
