+++ discovery/yahoosearch.py
@@ -2,6 +2,7 @@ from parsers import myparser
 import time
 import requests
 from discovery.constants import *
+from lib.core import *
 
 
 class search_yahoo:
@@ -19,7 +20,7 @@ class search_yahoo:
                + "\"&b=" + str(self.counter) + "&pz=10"
         headers = {
             'Host': self.hostname,
-            'User-agent': getUserAgent()
+            'User-agent': Core.get_user_agent()
         }
         h = requests.get(url=url, headers=headers)
         self.total_results += h.text
@@ -28,13 +29,13 @@ class search_yahoo:
         while self.counter <= self.limit and self.counter <= 1000:
             self.do_search()
             time.sleep(getDelay())
-            print("\tSearching " + str(self.counter) + " results...")
+            print(f'\tSearching  {self.counter} results...')
             self.counter += 10
 
     def get_emails(self):
-        rawres = myparser.parser(self.total_results, self.word)
+        rawres = myparser.Parser(self.total_results, self.word)
         return rawres.emails()
 
     def get_hostnames(self):
-        rawres = myparser.parser(self.total_results, self.word)
+        rawres = myparser.Parser(self.total_results, self.word)
         return rawres.hostnames()
