+++ requirements.txt
@@ -3,3 +3,4 @@ plotly>=3.4.2
 requests>=2.21.0
 texttable>=1.4.0
 shodan>=1.10.0
+pytest>=4.0.2
